requirements.txt
Flask
flask-cors
pyjwt
werkzeug